refactor(sequencer/feeds): Feed processing functionality extracted to libs
