+++ libs/feeds_processing/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "feeds_processing"
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+edition.workspace = true
+
+[dependencies]
+anomaly_detection = { path = "../../libs/anomaly_detection" }
+feed_registry = { path = "../feed_registry" }
+data_feeds = {path = "../data_feeds"}
+config = {path = "../config"}
+hex = "0.4.3"
+ringbuf = "0.4.7"
+tracing = { workspace = true }
+tokio = { workspace = true }
+alloy = { version = "0.8", features = [
+    "sol-types",
+    "contract",
+    "network",
+    "node-bindings",
+    "providers",
+    "rpc",
+    "rpc-client",
+    "rpc-types-eth",
+    "rpc-types-trace",
+    "signers",
+    "signer-keystore",
+    "signer-ledger",
+    "signer-mnemonic",
+    "signer-trezor",
+    "signer-yubihsm",
+    "transports",
+    "transport-http",
+    "transport-ipc",
+    "transport-ws",
+    "pubsub",
+] }
+alloy-primitives = "0.8.19"
+alloy-encode-packed = "0.1.1"
+anyhow = "1.0.89"
+serde_json = { workspace = true }
+serde = { workspace = true }
